pyproject.toml
[tool.poetry]
name = "weaviate-ml"
version = "0.1.0"
description = "Weaviate's challenge project"
authors = ["ugm2 <unaigaraymaestre@gmail.com>"]
readme = "README.md"
packages = [
    {include = "dataset", from = "src"},
    {include = "weaviate_ml", from = "src"},
    {include = "frontend", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.10"
weaviate-client = "^4.6.3"
transformers = "^4.41.1"
certifi = "^2024.2.2"
rich = "^13.7.1"
torch = "^2.3.0"
streamlit = "^1.35.0"
watchdog = "^4.0.1"
scikit-learn = "^1.5.0"
plotly = "^5.22.0"
streamlit-option-menu = "^0.3.13"


[tool.poetry.group.dataset.dependencies]
pandas = "^2.2.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
